fix errors for bevy 0.17
- Changed sound and music functions to accept `impl Into<String>` for flexibility
- Updated event handling to use `add_message` instead of `add_event` for core simulation
- Modified UI menu components to use `DespawnOnExit` for better resource management
- Adjusted button interaction handling to use `MessageWriter` for app exit
- Refined UI panel updates to trigger `CompileContextEvent` with structured data

+++ dominion_earth/src/main.rs
@@ -56,7 +56,7 @@ fn main() {
     app.add_plugins(DefaultPlugins.set(WindowPlugin {
         primary_window: Some(Window {
             title: window::TITLE.to_string(),
-            resolution: (window::DEFAULT_WIDTH, window::DEFAULT_HEIGHT).into(),
+            resolution: (window::DEFAULT_WIDTH as u32, window::DEFAULT_HEIGHT as u32).into(),
             mode: window_mode,
             ..default()
         }),
